Removing util_test.go. Moved to db_test.go
